Move MultiVerify to mg.core and add unit tests to get full coverage
